Convert interface into concrete impl
