+++ library/src/main/java/com/novoda/downloadmanager/lib/DownloadService.java
@@ -17,7 +17,6 @@
 package com.novoda.downloadmanager.lib;
 
 import android.app.AlarmManager;
-import android.app.Notification;
 import android.app.PendingIntent;
 import android.app.Service;
 import android.content.ContentResolver;
@@ -35,12 +34,10 @@ import android.os.IBinder;
 import android.os.Message;
 import android.os.Process;
 import android.support.annotation.NonNull;
-import android.support.v4.util.SimpleArrayMap;
 
 import com.novoda.downloadmanager.lib.logger.LLog;
 import com.novoda.downloadmanager.notifications.DownloadNotifier;
 import com.novoda.downloadmanager.notifications.DownloadNotifierFactory;
-import com.novoda.downloadmanager.notifications.NotificationTag;
 
 import java.io.File;
 import java.io.FileDescriptor;
@@ -53,8 +50,6 @@ import java.util.Map;
 import java.util.concurrent.ExecutorService;
 
 import static android.text.format.DateUtils.MINUTE_IN_MILLIS;
-import static com.novoda.downloadmanager.notifications.DownloadNotifier.NotificationsCreatedCallback;
-import static com.novoda.downloadmanager.notifications.DownloadNotifier.TYPE_ACTIVE;
 
 /**
  * Performs background downloads as requested by applications that use
@@ -93,24 +88,7 @@ public class DownloadService extends Service {
     private NetworkChecker networkChecker;
     private DestroyListener destroyListener;
 
-    private final NotificationsCreatedCallback notificationsCreatedCallback = new NotificationsCreatedCallback() {
-        @Override
-        public void onNotificationCreated(SimpleArrayMap<NotificationTag, Notification> taggedNotifications) {
-            boolean noActiveDownloads = true;
-
-            for (int i = 0; i < taggedNotifications.size(); i++) {
-                NotificationTag currentTag = taggedNotifications.keyAt(i);
-                if (currentTag.status() == TYPE_ACTIVE) {
-                    DownloadService.this.startForeground(currentTag.hashCode(), taggedNotifications.get(currentTag));
-                    noActiveDownloads = false;
-                    break;
-                }
-            }
-            if (noActiveDownloads) {
-                DownloadService.this.stopForeground(false);
-            }
-        }
-    };
+    private final NotificationCreatedListener notificationsCreatedCallback = new NotificationCreatedListener(this);
 
     /**
      * Receives notifications when the data in the content provider changes
@@ -481,5 +459,4 @@ public class DownloadService extends Service {
     protected void dump(FileDescriptor fd, @NonNull PrintWriter writer, String[] args) {
         LLog.e("I want to dump but nothing to dump into");
     }
-
 }
